+++ react-rsuite5/checkbox/index.js
@@ -16,19 +16,20 @@ const CheckboxRSuite = I18N(
     label,
     hint,
     value,
-    placeholder,
-    plaintext,
+
+    //placeholder,
+    //plaintext,
     tooltip = false,
-    disabled = false,
-    readOnly = false,
-    required = false,
+    //disabled = false,
+    //readOnly = false,
+    //required = false,
     indeterminate,
     error,
     onChange,
-    onBlur,
-    appearance,
-    locale,
-    format,
+    //onBlur,
+    //appearance,
+    //locale,
+    //format,
     ...rest
   }) => {
     const [isChecked, setIsChecked] = useState(value ?? null);
@@ -52,6 +53,8 @@ const CheckboxRSuite = I18N(
       [onChange, indeterminate, isChecked],
     );
 
+    console.log('oca rest', rest)
+
     return (
       <Form.Group data-lf-field-name={name} className="lt-control-checkbox">
         <RSuite5FieldControl errorMessage={_.isString(error) ? error : undefined }>
@@ -60,11 +63,8 @@ const CheckboxRSuite = I18N(
             name={name}
             checked={isChecked}
             onChange={handleChange}
-            readOnly={readOnly}
-            plaintext={plaintext}
-            onBlur={onBlur}
-            errorMessage={_.isString(error) ? error : undefined }
-            disabled={disabled}
+            //readOnly={readOnly}
+            //disabled={disabled}
             {...passRest(rest)}
           >
             {label}